requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
bcrypt==4.0.1
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
et-xmlfile==1.1.0
Flask==2.2.3
Flask-Cors==3.0.10
fonttools==4.39.4
frozenlist==1.3.3
fsspec==2023.5.0
gcsfs==2023.5.0
geographiclib==2.0
geopy==2.3.0
google-api-core==2.11.0
google-auth==2.18.0
google-auth-oauthlib==1.0.0
google-cloud-core==2.3.2
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
multidict==6.0.4
networkx==3.1
numpy==1.24.2
oauthlib==3.2.2
openpyxl==3.1.2
overpy==0.6
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
protobuf==4.23.0
psutil==5.9.5
psycopg2==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.0.9
PySimpleGUI==4.60.4
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.29.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
six==1.16.0
SQLAlchemy==2.0.13
tarjan==0.2.4
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
Werkzeug==2.2.3
yarl==1.9.2
